+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gaggia PID Controller Changelog
 
+## []
+### Added
+- Added python script to plot machine temperatures reported through serial interface 
+
 ## [1.0.0] - 2020-05-03
 ### Added
 - First release: